refactor: extract ios detail helpers
